anrcore: Don't crash on wild pointers.
This is a temporary fix to work around an issue with libdl and
thread-local storage. A bug in glibc can mix private calls to
__libc_memalign with overloaded calls to free that result in crashes.

On Lexmark devices, side effects of this bug are SIGABRTs in eSF JNI
code, with an error in anrcore.c about "free of wild pointer".

The broken glibc allocations are in elf/dl-tls.c, and this is tracked
upstream as BZ 9721, which was mistakenly marked as fixed. When that
bug is really closed, the previous behavior can likely be restored.