anrcore: Fix word boundary logic in find_treechunk().
The problem:
The logic in find_treechunk() was not proceeding to the next word in the
largemap[] when the index was the most significant bit (ex. 31), the bin at
that index was empty, and there were chunks in smaller bins.  In this
case index2bit(index+1) was returning 0, so mask_right() was not masking
the *bitmap.  Therefore, we were going into the if() instead of the else()
case and failing to find a block.

The solution:
If index2bit(index+1) returns a 0, that means the index is on a word
boundary (ex. 31, 63, etc.).  In this case, mask_right() will return
the entire *bitmap, because it is attempting to mask to the right of
bit 0.  Therefore, check the bit position before calling mask_right().
If it is 0, that means we want to go up to the next word in the largemap[].

Testing:
Expanded the largebins unit test to verify that find_treechunk() will
always discover the next largest bin.