gmalloc: add a EXPANDABLE_POOL_SIZE option
Rather than capping the memory for eSF at a specific value, we want to
be able to grow the pool size when additional bundles are installed. To
do so, initially initialize with a pool size of all available memory,
then shrink it to the requested initial value. The pool size can then be
grown at any time with _anr_core_expand().