anrcore.c: block can't be marked free before find_chunk()
_anr_core_realloc would clear the IN_USE bit of a chunk early in the
call, presumably to catch internal allocator bugs where we try to free
the chunk.  However, the chunk is not really free; it contains user data
and will not have a place in the largebin trees.  If find_chunk() goes
down the slab allocation path, we could potentially try to modify block.
In particular, if we chose the chunk immediately prior to block to use
for the new slab, internal_allocate_slab will see that next_chunk is not
IN_USE (even though it is) and try to join the two chunks.  This
can result in various sorts of badness (treating user data as pointers,
potentially), but has been observed to cause a crash when the largebin
for that size is empty (NULL):

    rb_remove_size (tree=0x0, tree@entry=0xb3c0a090, words=4102)
    unlink_chunk (block=0x1cbdecd0, self=0xb3c09000)
    internal_allocate_slab (self=self@entry=0xb3c09000, pool=0xabc5178)
    manage_slab (block=<optimized out>, self=0xb3c09000)
    find_binchunk (words=11, self=0xb3c09000)
    find_chunk (self=0xb3c09000, words=11)
    _anr_core_realloc (self=0xb3c09000, ptr=0x1cbdecd8, size=size@entry=80,

(Note that above, unlink_chunk is the same as the ptr we're trying to realloc).

In any case, block must be marked IN_USE during the call to
find_chunk().  Since the rest of the defragmenting realloc case doesn't
care about IN_USE at all, we can just wait until afterwards to clear the
bit.