gmalloc: do proper accounting for initial membroker reserve
The allocator has overhead.  If the mapping is very large, then the
allocator's overhead can be quite large.  I'm observing 512M of heap
overhead for a 2G mapping.  In the case that we want an initial pool
size of 32M, but we have a 2G mapping because the box has huge scads
of memory, and the internal dirty page bitmaps take up more than the
32M we reserve for the initial pool size.

In any case, expandable pool size or not, we have to reserve not only
the pool size but also the overhead, or we will be using unbudgeted
memory.  Rework the init code to shrink and expand to find overhead
and do all the things properly.