unlink_chunk: verify that block is actually on a list
Calling unlink_chunk with a block that's not on a list can corrupt our
data structures.  In particular, when both link.next and link.prev are
NULL, they will be equal to each other, which this code interprets
as "this was the last thing in the bin where I would expect this object
to live."  The code then empties out the bin, even though this object
wasn't even in the list!  This can cause debug assertions later when we
go looking for one of the objects that used to be in the list...